Complete A-CodingStandard. Make changes for program to align better with given coding standards
