+++ src/main/java/tasks/Deadline.java
@@ -7,9 +7,9 @@ import java.time.format.DateTimeFormatter;
  * Represents a task with a deadline that must be completed before a specific date/time.
  */
 public class Deadline extends Task {
-    private LocalDateTime dueDate;
     private static final DateTimeFormatter OUTPUT_FORMATTER = DateTimeFormatter.ofPattern("EEEE, MMM dd yyyy, h:mma");
     private static final DateTimeFormatter TIME_FORMATTER = DateTimeFormatter.ofPattern("h:mma");
+    private final LocalDateTime dueDate;
 
     /**
      * Constructs a new Deadline task.
